MVVM is working, reusing Rv + SettingsFragment with drag and drop Rvs(need to fix)

+++ app/src/main/java/com/example/antip/App.kt
@@ -3,4 +3,4 @@ package com.example.antip
 import android.graphics.drawable.Drawable
 
 
-data class App(val image:Drawable,val packageName:String,val scores:Long)
+data class App(val image:Drawable,val packageName:String,val scores:Int)
